move map component into its own file
